# docs/swagger.yaml
basePath: /
definitions:
  controllers.adminlogin:
    properties:
      adminMail:
        type: string
      adminPassword:
        type: string
    type: object
  controllers.otpvalidation:
    properties:
      otp:
        type: string
    type: object
  controllers.signupdata:
    properties:
      email:
        type: string
      gender:
        type: string
      mobile:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  controllers.userlogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  handlers.addcoupon:
    properties:
      code:
        type: string
      condition:
        type: integer
      discount:
        type: number
      validfrom:
        type: string
      validto:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ecom application
  title: swagger documentation for the application
  version: "1.0"
paths:
...
  /admin/category:
    get:
      consumes:
      - application/json
      description: Admin  can list categories
      produces:
      - application/json
      responses:
        "200":
          description: Added Category
          schema:
            type: json
        "401":
          description: Failed to add Category
          schema:
            type: json
      summary: List Category
      tags:
      - Admin-CategoryManagement
  /admin/category/delete/{ID}:
    delete:
      consumes:
      - application/json
      description: Admin  can Delete category by selecting the category
      parameters:
      - description: Category ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted Category
          schema:
            type: json
        "401":
          description: Category not Found or Failed to Delete Category
          schema:
            type: json
      summary: Delete Category
      tags:
      - Admin-CategoryManagement
  /admin/category/edit/{ID}:
    post:
      consumes:
      - application/json
      description: Admin  can Edit category
      parameters:
      - description: Category ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Edited Category
          schema:
            type: json
        "401":
          description: Category not Found or Failed to Edit Category
          schema:
            type: json
      summary: Edit Category
      tags:
      - Admin-CategoryManagement
  /admin/coupon:
    get:
      consumes:
      - application/json
      description: Admin  can View all the coupons
      produces:
      - application/json
      responses:
        "200":
          description: Listed all Coupons
          schema:
            type: json
        "401":
          description: Coupon not Found or Failed to List Coupon
          schema:
            type: json
      summary: List Coupon
      tags:
      - Admin-CouponManagement
    post:
      consumes:
      - application/json
      description: Admin  can Add new coupons
      parameters:
      - description: Coupon details
        in: body
        name: coupon
        required: true
        schema:
          $ref: '#/definitions/handlers.addcoupon'
      produces:
      - application/json
      responses:
        "200":
          description: Added new Coupon
          schema:
            type: json
        "401":
          description: Coupo not Found or Failed to Add New Coupon
          schema:
            type: json
      summary: Add Coupon
      tags:
      - Admin-CouponManagement
  /admin/coupon/{ID}:
    delete:
      consumes:
      - application/json
      description: Admin  can Delete coupons with selected Coupon ID
      parameters:
      - description: Coupon ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted Coupons
          schema:
            type: json
        "401":
          description: Coupon not Found or Failed to Delete Coupon
          schema:
            type: json
      summary: Delete Coupon
      tags:
      - Admin-CouponManagement
  /admin/login:
    post:
      consumes:
      - application/json
      description: Logs in the Admin with email and password, returns a token to be
        used for authentication
      parameters:
      - description: Login Data
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/controllers.adminlogin'
      produces:
      - application/json
      responses:
        "200":
          description: Login Successful".
          schema:
            type: json
        "401":
          description: Unauthorized.
          schema:
            type: json
      summary: Login as an admin
      tags:
      - Admin-Auth
  /admin/logout:
    get:
      consumes:
      - application/json
      description: Logs out as an admin, clear the token used for authentication
      produces:
      - application/json
      responses:
        "200":
          description: Logout Successful".
          schema:
            type: json
        "401":
          description: Unauthorized.
          schema:
            type: json
      summary: Logout as an admin
      tags:
      - Admin-Auth
...
  /user/logout:
    get:
      consumes:
      - application/json
      description: Logs out as an user, clear the token used for authentication
      produces:
      - application/json
      responses:
        "200":
          description: Logout Successful".
          schema:
            type: json
        "401":
          description: Unauthorized.
          schema:
            type: json
      summary: Logout as an User
      tags:
      - User-Auth
...
  /user/registration:
    post:
      consumes:
      - application/json
      description: Collecting the New User data and Send  an OTP to verify Email ID
      parameters:
      - description: SignUp Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.signupdata'
      produces:
      - application/json
      responses:
        "200":
          description: ' Successfuly sent OTP".'
          schema:
            type: json
        "401":
          description: Failed to sent OTP.
          schema:
            type: json
      summary: User Signup
      tags:
      - User-Auth
  /user/registration/otp:
    post:
      consumes:
      - application/json
      description: Validating OTP
      parameters:
      - description: Otp check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.otpvalidation'
      produces:
      - application/json
      responses:
        "200":
          description: ' Successfully signed up'
          schema:
            type: json
        "401":
          description: Failed to Signup
          schema:
            type: json
      summary: OTP Check
      tags:
      - User-Auth
  /user/registration/resendotp:
    post:
      consumes:
      - application/json
      description: This API is used for resending the OTP
      produces:
      - application/json
      responses:
        "200":
          description: ' Resent OTP'
          schema:
            type: json
        "401":
          description: Failed to send OTP
          schema:
            type: json
      summary: Resend OTP
      tags:
      - User-Auth
...
  /user/signin:
    post:
      consumes:
      - application/json
      description: Logs in the User with email and password, returns a token to be
        used for authentication
      parameters:
      - description: Login Data
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/controllers.userlogin'
      produces:
      - application/json
      responses:
        "200":
          description: Login Successful".
          schema:
            type: json
        "401":
          description: Unauthorized.
          schema:
            type: json
      summary: Login as an User
      tags:
      - User-Auth
swagger: "2.0"